Use single CustomPropertiesBackend for old and new dav endpoints
