+++ apps/dav/lib/Connector/Sabre/ServerFactory.php
@@ -29,6 +29,7 @@
 namespace OCA\DAV\Connector\Sabre;
 
 use OCA\DAV\Files\BrowserErrorPagePlugin;
+use OCA\DAV\Files\CustomPropertiesBackend;
 use OCP\Files\Mount\IMountManager;
 use OCP\IConfig;
 use OCP\IDBConnection;
@@ -180,7 +181,7 @@ class ServerFactory {
 				// custom properties plugin must be the last one
 				$server->addPlugin(
 					new \Sabre\DAV\PropertyStorage\Plugin(
-						new \OCA\DAV\Connector\Sabre\CustomPropertiesBackend(
+						new CustomPropertiesBackend(
 							$objectTree,
 							$this->databaseConnection,
 							$this->userSession->getUser()
