+++ apps/dav/tests/unit/Files/CustomPropertiesBackendTest.php
@@ -20,7 +20,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>
  *
  */
-namespace OCA\DAV\Tests\unit\Connector\Sabre;
+namespace OCA\DAV\Tests\unit\Files;
 
 /**
  * Copyright (c) 2015 Vincent Petry <pvince81@owncloud.com>
@@ -34,7 +34,7 @@ namespace OCA\DAV\Tests\unit\Connector\Sabre;
  *
  * @group DB
  *
- * @package OCA\DAV\Tests\unit\Connector\Sabre
+ * @package OCA\DAV\Tests\unit\Files
  */
 class CustomPropertiesBackendTest extends \Test\TestCase {
 
@@ -49,7 +49,7 @@ class CustomPropertiesBackendTest extends \Test\TestCase {
 	private $tree;
 
 	/**
-	 * @var \OCA\DAV\Connector\Sabre\CustomPropertiesBackend
+	 * @var \OCA\DAV\Files\CustomPropertiesBackend
 	 */
 	private $plugin;
 
@@ -75,7 +75,7 @@ class CustomPropertiesBackendTest extends \Test\TestCase {
 			->method('getUID')
 			->will($this->returnValue($userId));
 
-		$this->plugin = new \OCA\DAV\Connector\Sabre\CustomPropertiesBackend(
+		$this->plugin = new \OCA\DAV\Files\CustomPropertiesBackend(
 			$this->tree,
 			\OC::$server->getDatabaseConnection(),
 			$this->user
